fx(backuping): adds check current woking directory before cleaning
because if user has no right to change directory to specified, scripts will work in current directory and at the end it will try clean current directory

+++ backuping/db_autobackup.sh
@@ -179,8 +179,10 @@ else
     fi
 fi
 
-echo "[$(date +%Y-%m-%d\ %H:%M:%S)] clean $work_dir"
-rm ./*
+if [[ "$(pwd)" == "$work_dir" ]]; then
+    echo "[$(date +%Y-%m-%d\ %H:%M:%S)] clean $work_dir"
+    rm ./*
+fi
 
 echo "[$(date +%Y-%m-%d\ %H:%M:%S)] restore working directory"
 popd
